refactor: update news controller

+++ controllers/news.js
@@ -1,5 +1,42 @@
 const News = require("../models/news");
-const User = require("../models/user");
+const Editor = require("../models/editor");
+
+exports.createNews = async (req, res, next) => {
+  try {
+    const editorID = req.userId;
+
+    const news = await News({
+      title: req.body.title,
+      content: req.body.content,
+      editorID: editorID,
+    }).save();
+
+    const editor = await Editor.findByIdAndUpdate(editorID, {
+      $push: { news },
+    });
+
+    res.status(201).json({
+      message: "News created.",
+      news: {
+        id: news.id,
+        title: news.title,
+        content: news.content,
+        createdAt: news.createdAt,
+        editor: {
+          id: editor._id,
+          name: editor.name,
+          surname: editor.surname,
+        },
+      },
+    });
+  } catch (err) {
+    if (!err.statusCode) {
+      err.statusCode = 500;
+    }
+
+    next(err);
+  }
+};
 
 exports.getNews = async (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -11,23 +48,31 @@ exports.getNews = async (req, res, next) => {
       .skip((currentPage - 1) * perPage)
       .limit(perPage);
 
+    const output = news.map((item) => {
+      return {
+        id: item.id,
+        title: item.title,
+        content: item.content,
+        createdAt: item.createdAt,
+      };
+    });
+
     res.status(200).json({
-      news: news,
+      news: output,
       countOfNews: countOfNews,
     });
   } catch (err) {
     if (!err.statusCode) {
       err.statusCode = 500;
     }
+
     next(err);
   }
 };
 
 exports.getNewsById = async (req, res, next) => {
-  const id = req.params.id;
-
   try {
-    const news = await News.findById(id);
+    const news = await News.findById(req.params.id);
 
     if (!news) {
       const error = new Error("Could not find news.");
@@ -35,52 +80,33 @@ exports.getNewsById = async (req, res, next) => {
       throw error;
     }
 
+    const editor = await Editor.findById(news.editorID);
+
     res.status(200).json({
-      news: news,
+      news: {
+        id: news.id,
+        title: news.title,
+        content: news.content,
+        createdAt: news.createdAt,
+        editor: {
+          id: editor._id,
+          name: editor.name,
+          surname: editor.surname,
+        },
+      },
     });
   } catch (err) {
     if (!err.statusCode) {
       err.statusCode = 500;
     }
-    next(err);
-  }
-};
 
-exports.createNews = async (req, res, next) => {
-  const title = req.body.title;
-  const content = req.body.content;
-
-  const news = News({
-    title: title,
-    content: content,
-    editor: req.userId,
-  });
-
-  try {
-    await news.save();
-
-    const user = await User.findById(req.userId);
-    user.news.push(news);
-
-    await user.save();
-
-    res.status(201).json({
-      message: "News created.",
-      news: news,
-    });
-  } catch (err) {
-    if (!err.statusCode) {
-      err.statusCode = 500;
-    }
     next(err);
   }
 };
 
 exports.updateNews = async (req, res, next) => {
   const id = req.params.id;
-
-  const title = req.body.title;
-  const content = req.body.content;
+  const editorID = req.userId;
 
   try {
     const news = await News.findById(id);
@@ -91,25 +117,37 @@ exports.updateNews = async (req, res, next) => {
       throw error;
     }
 
-    news.title = title;
-    news.content = content;
+    if (news.editorID.toString() !== editorID) {
+      const error = new Error("Not authorized!'");
+      error.statusCode = 403;
+      throw error;
+    }
 
-    const result = await news.save();
+    news.title = req.body.title;
+    news.content = req.body.content;
+
+    const updatedNews = await news.save();
 
     res.status(200).json({
-      message: "News updated",
-      news: result,
+      message: "News updated.",
+      news: {
+        title: updatedNews.title,
+        content: updatedNews.content,
+        updatedAt: updatedNews.updatedAt,
+      },
     });
   } catch (err) {
     if (!err.statusCode) {
       err.statusCode = 500;
     }
+
     next(err);
   }
 };
 
 exports.deleteNews = async (req, res, next) => {
   const id = req.params.id;
+  const editorID = req.userId;
 
   try {
     const news = await News.findById(id);
@@ -120,11 +158,17 @@ exports.deleteNews = async (req, res, next) => {
       throw error;
     }
 
+    if (news.editorID.toString() !== editorID) {
+      const error = new Error("Not authorized!'");
+      error.statusCode = 403;
+      throw error;
+    }
+
     await News.findByIdAndRemove(id);
 
-    const user = await User.findById(req.userId);
-    user.news.pull(id);
-    await user.save();
+    await Editor.findByIdAndUpdate(news.editorID, {
+      $pull: { news: id },
+    });
 
     res.status(200).json({
       message: "Deleted news.",
@@ -133,6 +177,7 @@ exports.deleteNews = async (req, res, next) => {
     if (!err.statusCode) {
       err.statusCode = 500;
     }
+
     next(err);
   }
 };
